+++ NEWS.md
@@ -1,4 +1,6 @@
 # ILSAstats 0.4.5
+- Added `repmeanfast()`. - Missing option for PVs and by.
+- Added creation of replicate weights by index.
 - Added extensive documentation.
 - Added support for TIMSS Longitudinal 2023.
 - Added `autoILSA()`.